Bugfix: It should be Complex128 and not Complex64 in order to get 64-bit precision (see previous commit).

+++ src/NEPTypes.jl
@@ -194,6 +194,6 @@ module NEPTypes
     end
 
 
-    interpolate(nep::NEP, intpoints::Array) = interpolate(Complex64, nep, intpoints)
+    interpolate(nep::NEP, intpoints::Array) = interpolate(Complex128, nep, intpoints)
 
 end
